Move everything to a compat folder
