+++ music/demos/melody_controls.ts
@@ -19,7 +19,7 @@ import * as tf from '@tensorflow/tfjs';
 
 import * as mm from '../src/index';
 import {INoteSequence} from '../src/index';
-import * as timer from '../src/core/timer';
+import * as timer from '../src/core/compat/timer';
 
 import {CHECKPOINTS_DIR, MEL_TEAPOT, MEL_TWINKLE} from './common';
 import {writeMemory, writeNoteSeqs, writeTimer} from './common';
