+++ music/src/coconet/model.ts
@@ -20,10 +20,10 @@
  */
 import * as tf from '@tensorflow/tfjs-core';
 
-import {fetch} from '../core/fetch';
+import {fetch} from '../core/compat/fetch';
 import * as logging from '../core/logging';
 import * as sequences from '../core/sequences';
-import * as timer from '../core/timer';
+import * as timer from '../core/compat/timer';
 import {INoteSequence} from '../protobuf/index';
 
 import {IS_IOS, NUM_PITCHES, pianorollToSequence, sequenceToPianoroll} from './coconet_utils';
