+++ music/src/core/compat/fetch.ts
