+++ music/src/core/compat/fetch_node.ts
