+++ music/src/core/compat/timer.ts
