+++ music/src/core/compat/timer_node.ts
