+++ music/src/music_rnn/model.ts
@@ -25,10 +25,10 @@ import * as tf from '@tensorflow/tfjs';
 import * as aux_inputs from '../core/aux_inputs';
 import * as chords from '../core/chords';
 import * as data from '../core/data';
-import {fetch} from '../core/fetch';
+import {fetch} from '../core/compat/fetch';
 import * as logging from '../core/logging';
 import * as sequences from '../core/sequences';
-import * as timer from '../core/timer';
+import * as timer from '../core/compat/timer';
 import {INoteSequence} from '../protobuf/index';
 
 import {ATTENTION_PREFIX, AttentionWrapper} from './attention';
