+++ music/src/music_vae/model.ts
@@ -25,9 +25,9 @@ import * as tf from '@tensorflow/tfjs';
 import * as chords from '../core/chords';
 import * as constants from '../core/constants';
 import * as data from '../core/data';
-import {fetch} from '../core/fetch';
+import {fetch} from '../core/compat/fetch';
 import * as logging from '../core/logging';
-import * as timer from '../core/timer';
+import * as timer from '../core/compat/timer';
 import {INoteSequence} from '../protobuf/index';
 
 /**
