+++ music/src/piano_genie/model.ts
@@ -18,7 +18,7 @@
 
 import * as tf from '@tensorflow/tfjs';
 
-import {fetch} from '../core/fetch';
+import {fetch} from '../core/compat/fetch';
 
 /**
  * Constants.
