+++ music/src/transcription/model.ts
@@ -23,9 +23,9 @@
 import * as tf from '@tensorflow/tfjs';
 
 import {loadAudioFromFile, loadAudioFromUrl} from '../core/audio_utils';
-import {fetch} from '../core/fetch';
+import {fetch} from '../core/compat/fetch';
 import * as logging from '../core/logging';
-import * as timer from '../core/timer';
+import * as timer from '../core/compat/timer';
 import {INoteSequence} from '../protobuf/index';
 
 import {preprocessAudio} from './audio_utils';
